pyproject.toml
[tool.poetry]
name = "ha-appdaemon"
version = "0.10.1"
description = ""
authors = ["Will Garside <worgarside@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
wg-utilities = {extras = ["clients"], version = "^5.15.1"}
pygithub = "^2.3.0"
gitpython = "^3.1.43"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
appdaemon = "^4.4.2"

[tool.mypy]
plugins = "pydantic.mypy"

exclude = ["_local_sandbox"]
follow_imports = "normal"
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true

warn_return_any = true
warn_unused_ignores = true
warn_unused_configs = true
warn_redundant_casts = true

no_implicit_optional = true
strict_optional = true

strict_equality = true
explicit_package_bases = true
show_error_codes = true
enable_error_code = ["truthy-bool", "unused-awaitable"]

[tool.codespell]
skip = '*.json,*.csv,*.xml,*.lock'
ignore-words-list = "hass,ags"

[tool.ruff]
fix = true
line-length = 90
lint.select = [
  "A",    # https://docs.astral.sh/ruff/rules/#flake8-builtins-a
  "ARG",  # https://docs.astral.sh/ruff/rules/#flake8-unused-arguments-arg
  "B",    # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
  "C4",   # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
  "COM",  # https://docs.astral.sh/ruff/rules/#flake8-commas-com
  "D",    # https://docs.astral.sh/ruff/rules/#pydocstyle-d
  "E",    # https://docs.astral.sh/ruff/rules/#error-e
  "ERA",  # https://docs.astral.sh/ruff/rules/#eradicate-era
  "F",    # https://docs.astral.sh/ruff/rules/#pyflakes-f
  "FA",   # https://docs.astral.sh/ruff/rules/#flake8-future-annotations-fa
  "FBT",  # https://docs.astral.sh/ruff/rules/#flake8-boolean-trap-fbt
  "FLY",  # https://docs.astral.sh/ruff/rules/#flynt-fly
  "FURB", # https://docs.astral.sh/ruff/rules/#refurb-furb
  "G",    # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
  "I",    # https://docs.astral.sh/ruff/rules/#isort-i
  "ICN",  # https://docs.astral.sh/ruff/rules/#flake8-import-conventions-icn
  "INT",  # https://docs.astral.sh/ruff/rules/#flake8-gettext-int
  "ISC",  # https://docs.astral.sh/ruff/rules/#flake8-implicit-str-concat-isc
  "LOG",  # https://docs.astral.sh/ruff/rules/#flake8-logging-log
  "N",    # https://docs.astral.sh/ruff/rules/#pep8-naming-n
  "PERF", # https://docs.astral.sh/ruff/rules/#perflint-perf
  "PGH",  # https://docs.astral.sh/ruff/rules/#pygrep-hooks-pgh
  "PIE",  # https://docs.astral.sh/ruff/rules/#flake8-pie-pie
  "PL",   # https://docs.astral.sh/ruff/rules/#pylint-pl
  "PGH",  # https://docs.astral.sh/ruff/rules/#pygrep-hooks-pgh
  "PT",   # https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
  "PTH",  # https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
  "Q",    # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
  "RET",  # https://docs.astral.sh/ruff/rules/#flake8-return-ret
  "RSE",  # https://docs.astral.sh/ruff/rules/#flake8-raise-rse
  "RUF",  # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
  "S",    # https://docs.astral.sh/ruff/rules/#flake8-bandit-s
  "SIM",  # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
  "TCH",  # https://docs.astral.sh/ruff/rules/#flake8-type-checking-tch
  "TID",  # https://docs.astral.sh/ruff/rules/#flake8-tidy-imports-tid
  "TRIO", # https://docs.astral.sh/ruff/rules/#flake8-trio-trio
  "TRY",  # https://docs.astral.sh/ruff/rules/#tryceratops-try
  "UP",   # https://docs.astral.sh/ruff/rules/#pyupgrade-up
  "W",    # https://docs.astral.sh/ruff/rules/#warning-w
]
lint.ignore = [
  "D107",   # https://docs.astral.sh/ruff/rules/undocumented-public-init/
  "S314",   # https://docs.astral.sh/ruff/rules/suspicious-xml-element-tree-usage/
  "TRY003", # https://docs.astral.sh/ruff/rules/raise-vanilla-args/
  "ISC001", # https://docs.astral.sh/ruff/rules/single-line-implicit-string-concatenation/
]

[tool.ruff.lint.pycodestyle]
max-line-length = 115

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
